.perl
Output
This service returns data as a set of comma-separated fields. The individual fields are not escaped or quoted, but they will never contain a comma.

All strings are returned as ASCII.

Name	Type (length)
proxyScore	decimal	A score from 0.00-4.00 indicating the likelihood that the user's IP address is an anonymous proxy, open proxy, or VPN.
proxyScore	Likelihood of fraud
0.5	15%
1.0	30%
2.0	60%
3.0+	90% A proxyScore of 0.00 will be returned for a corporate proxy or private IP and an empty string will be returned for an invalid IP.
err	enum	
If there was an error or warning with this request, this field contains an error code string.

The possible error codes are:

LICENSE_REQUIRED – you must provide a license key.
MAX_REQUESTS_REACHED – this error will be returned if your account is out of queries or if an invalid license key is provided.
Client Code Examples
Below are some sample clients for this web service.